Add 16bit support for xtensa max and average pooling kernels
